requirements.txt

Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
groq==0.9.0
python-dotenv==1.0.0
pathlib2==2.3.7.post1; python_version < '3.4'
black==23.12.1
flake8==6.1.0



